+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -154,6 +154,11 @@ New checks
 
   FIXME: add release notes.
 
+- New :doc:`openmp-overlapping-dependency-storage
+  <clang-tidy/checks/openmp/overlapping-dependency-storage>` check.
+
+  FIXME: add release notes.
+
 - New :doc:`openmp-recursive-task-without-task-creation-condition
   <clang-tidy/checks/openmp/recursive-task-without-task-creation-condition>` check.